LayoutTests/imported/w3c/web-platform-tests/selection/fire-selectionchange-event-on-textcontrol-element-on-pressing-backspace-expected.txt



FAIL selectionchange event fired on an input element assert_equals: selectionStart after pressing backspace key expected 4 but got 5
